requirements.txt
streamlit
pandas
folium
streamlit-folium
python-dotenv
requests
streamlit-js-eval
